+++ android/src/main/kotlin/cat/durban/sergio/flutter_meta_appads_sdk/helper/CurrencyFinder.kt
@@ -1,3 +1,5 @@
+package cat.durban.sergio.flutter_meta_appads_sdk.helper
+
 import java.util.Currency
 import java.util.Optional